Refactor identifier case environment setup
